added second degree solving, need to switch first degree and Term class to gmp and it should be done

+++ Term.hpp
@@ -2,22 +2,23 @@
 # define TERM_HPP
 # include <string>
 # include <vector>
+# include <gmpxx.h>
 
 class Term
 {
 	public:
 		Term(std::string newTerm, bool ispositive);
-		Term(int exponent, long value);
+		Term(int exponent, int value);
 		~Term();
 		class	EmptyValue : public std::exception {const char *what() const throw();};
 		void	ParseAndAdd(std::string val, bool ispositive);
 		Term	&operator+=(const Term& rhs);
 		int		getExponent() const;
-		long	getValue() const;
+		int	getValue() const;
 		void	reverseValue();
 	private:
 		int	_exponent = -1;
-		long	_value = 0;
+		int	_value = 0;
 		Term();
 
 };
